Move silx.gui.test.utils to silx.gui.utils.testutils
