+++ silx/gui/data/test/test_arraywidget.py
@@ -34,7 +34,7 @@ import numpy
 
 from silx.gui import qt
 from silx.gui.data import ArrayTableWidget
-from silx.gui.test.utils import TestCaseQt
+from silx.gui.utils.testutils import TestCaseQt
 
 try:
     import h5py
