+++ silx/gui/plot/test/testAlphaSlider.py
@@ -33,7 +33,7 @@ import numpy
 import unittest
 
 from silx.gui import qt
-from silx.gui.test.utils import TestCaseQt
+from silx.gui.utils.testutils import TestCaseQt
 from silx.gui.plot import PlotWidget
 from silx.gui.plot import AlphaSlider
