+++ silx/gui/plot/test/testImageView.py
@@ -33,7 +33,7 @@ import unittest
 import numpy
 
 from silx.gui import qt
-from silx.gui.test.utils import TestCaseQt
+from silx.gui.utils.testutils import TestCaseQt
 
 from silx.gui.plot import ImageView
 from silx.gui.colors import Colormap
