+++ silx/gui/plot/test/testProfile.py
@@ -32,7 +32,7 @@ import numpy
 import unittest
 
 from silx.utils.testutils import ParametricTestCase
-from silx.gui.test.utils import (
+from silx.gui.utils.testutils import (
     TestCaseQt, getQToolButtonFromAction)
 from silx.gui import qt
 from silx.gui.plot import PlotWindow, Plot1D, Plot2D, Profile
