+++ silx/gui/plot/tools/test/testCurveLegendsWidget.py
@@ -31,7 +31,7 @@ import unittest
 
 from silx.gui import qt
 from silx.utils.testutils import ParametricTestCase
-from silx.gui.test.utils import TestCaseQt
+from silx.gui.utils.testutils import TestCaseQt
 from silx.gui.plot import PlotWindow
 from silx.gui.plot.tools import CurveLegendsWidget
