+++ silx/gui/plot3d/test/testScalarFieldView.py
@@ -34,7 +34,7 @@ import unittest
 import numpy
 
 from silx.utils.testutils import ParametricTestCase
-from silx.gui.test.utils import TestCaseQt
+from silx.gui.utils.testutils import TestCaseQt
 from silx.gui import qt
 
 from silx.gui.plot3d.ScalarFieldView import ScalarFieldView
