+++ silx/gui/test/test_console.py
@@ -33,7 +33,7 @@ __date__ = "05/12/2016"
 
 import unittest
 
-from silx.gui.test.utils import TestCaseQt
+from silx.gui.utils.testutils import TestCaseQt
 
 from silx.gui import qt
 try:
