+++ silx/gui/utils/test/test_image.py
@@ -33,7 +33,7 @@ import unittest
 
 from silx.gui import qt
 from silx.utils.testutils import ParametricTestCase
-from silx.gui.test.utils import TestCaseQt
+from silx.gui.utils.testutils import TestCaseQt
 from silx.gui.utils.image import convertArrayToQImage, convertQImageToArray
