+++ silx/gui/widgets/test/test_printpreview.py
@@ -30,7 +30,7 @@ __date__ = "19/07/2017"
 
 
 import unittest
-from silx.gui.test.utils import TestCaseQt
+from silx.gui.utils.testutils import TestCaseQt
 from silx.gui.widgets.PrintPreview import PrintPreviewDialog
 from silx.gui import qt
