+++ silx/gui/widgets/test/test_tablewidget.py
@@ -30,7 +30,7 @@ __date__ = "05/12/2016"
 
 
 import unittest
-from silx.gui.test.utils import TestCaseQt
+from silx.gui.utils.testutils import TestCaseQt
 from silx.gui.widgets.TableWidget import TableWidget
