+++ silx/gui/widgets/test/test_threadpoolpushbutton.py
@@ -32,8 +32,8 @@ __date__ = "17/01/2018"
 import unittest
 import time
 from silx.gui import qt
-from silx.gui.test.utils import TestCaseQt
-from silx.gui.test.utils import SignalListener
+from silx.gui.utils.testutils import TestCaseQt
+from silx.gui.utils.testutils import SignalListener
 from silx.gui.widgets.ThreadPoolPushButton import ThreadPoolPushButton
 from silx.utils.testutils import TestLogging
