+++ silx/sx/test/test_sx.py
@@ -24,7 +24,7 @@
 # ###########################################################################*/
 __authors__ = ["T. Vincent", "P. Knobel"]
 __license__ = "MIT"
-__date__ = "24/04/2018"
+__date__ = "05/10/2018"
 
 
 import logging
@@ -36,7 +36,7 @@ from silx.test.utils import test_options
 
 from silx.gui import qt
 # load TestCaseQt before sx
-from silx.gui.test.utils import TestCaseQt
+from silx.gui.utils.testutils import TestCaseQt
 from silx.gui.colors import rgba
 from silx import sx
